+++ apps/examples/src/App.tsx
@@ -1,9 +1,9 @@
-import { PlayCoreAsciiSettings, PlayCoreAscii } from "react-ascii-play";
+import { AsciiRendererSettings, ReactAsciiPlay } from "react-ascii-play";
 import "./css/font.css";
 import "./App.css";
 import torusProgram from "./examples/torus";
 
-const settings: PlayCoreAsciiSettings = {
+const settings: AsciiRendererSettings = {
   renderer: "canvas",
   fps: 60,
   fontFamily: "Simple Console, monospace",
@@ -13,7 +13,7 @@ function App() {
   return (
     <div className="app">
       <div className="playground-container">
-        <PlayCoreAscii
+        <ReactAsciiPlay
           className="simple-console-font"
           program={torusProgram}
           settings={settings}